fix(#66): footer static on mobile
